# backend/routes/spec/api-docs.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Digital Realty Website Backend API
  description: |
    Provides the backend/bridge for the digital realty website
  contact:
    name: Propane Studios
servers:
  - url: /api/v1
    description: Version 1 of this server
security:
  - basicAuth: []
paths:
...
  /project:
    get:
      summary: List all projects
      operationId: listProjects
      tags:
        - Project
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Limit"
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
        - name: title
          in: query
          description: Project Title
          required: false
          schema:
            type: string
        - name: archived
          in: query
          description: Get archived or unarchived projects
          required: false
          schema:
            $ref: "#/components/schemas/ArchivedFields"
        - name: useCase
          in: query
          description: array of ids of the use cases to search for
          required: false
          schema:
            type: array
            items:
              type: string
              example: bc2760ec-dfb7-40df-a360-7ad0f87a5b3b
        - name: order
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderFields'

              
      responses:
        '200':
          description: A paged array of projects that the user has access to
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  total:
                    $ref: "#/components/schemas/Total"
                  numPages:
                    $ref: "#/components/schemas/NumPages"
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
    post:
      tags:
        - Project
      operationId: createProject
      summary: Creates a project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostProject"
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessProject"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /project/{id}:
    get:
      tags:
        - Project
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/IdFields'
          required: true
      responses:
        '200':
          description: The requested project
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    type: object
                    $ref: "#/components/schemas/Project"
                
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Project
      summary: used to update data of a project
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/PutProject"

      responses:
        '200':
          description: Successfully updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessProject"

        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Project
      summary: Set a project to archived

      responses:
        '200':
          description: Successfully archived project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessProject"

        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Total:
      type: number
      description: total number of records
    NumPages:
      type: number
      description: total number of pages given the limit
    Limit:
      type: number
      description: Total number of records per page (max 100)
    Page:
      description: Which page to start on (relative to limit)
      type: integer
      format: int32
    PhoneFields:
      type: string
      description: User's phone number
      maxLength: 20
      pattern: '/^\s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d+))?\s*$/'
      example: 1-555-555-5555
    EmailFields:
      type: string
      description: User's email address
      pattern: /^\w+([.-]?\w+)+([+]?(\d|\w))*@\w+([.-]?\w+)*(\.\w{2,3})+$/
      example: 'admin@pdxmodeler.com'
    OrderFields:
      description: Fields (i.e. archived) followed by their sort order (ASC/DESC)
      type: array
      items:
        type: array
        items: 
          type: string
      example: [['archived', 'DESC']]
    UserFirstNameField:
      type: string
      description: First Name of the User
    UserLastNameField:
      type: string
      description: Last Name of the User
    CompanyNameField:
      type: string
      description: name of the company
    UsernameField:
      type: string
      maxLength: 60
      description: user's username (for login)
    UserRoleField:
      type: string
      description: User's role ID
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
...
    Company:
      required:
        - name
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        name:
          $ref: '#/components/schemas/CompanyNameField'
        createdBy:
          $ref: '#/components/schemas/CreatedBy'
...
    User:
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - username
        - role
        - archived
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        firstName:
          $ref: '#/components/schemas/UserFirstNameField'
        lastName:
          $ref: '#/components/schemas/UserLastNameField'
        email:
          $ref: "#/components/schemas/EmailFields"
        phone:
          $ref: "#/components/schemas/PhoneFields"
        username:
          $ref: '#/components/schemas/UsernameField'
        role:
          $ref: '#/components/schemas/UserRoleField'
        archived:
          $ref: '#/components/schemas/ArchivedFields'
...
    InvalidRequest:
      required:
        - error
      properties:
        error:
          type: string
          description: Contains a message detailing why the request was invalid
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: contains a message about the failure
...
    ProjectTitle:
      type: string
      description: Title of the project
      example: My Project title
    ProjectCompanyId:
      type: string
      description: Id for the parent company
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
    ProjectUseCases:
      type: array
      description: Use case ids for the project
      items:
        type: string
        example: b0d0ab14-4069-48df-9258-86647a0d7d95
    CreatedBy:
      type: string
      description: id for creating user
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
    IdFields:
      type: string
      description: Object id (UUID)
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
...
    ArchivedFields:
      type: boolean
      description: Whether or not this is archived
      example: false
    AssociatedUserFields:
      type: array
      description: Array of ids for associated users
      items:
        $ref: '#/components/schemas/IdFields'
    CreatedAtFields:
      type: string
      description: date the record was created
      example: '2022-09-16T13:00:00.000Z'
    UpdatedAtFields:
      type: string
      description: date the record was updated
      example: '2022-09-16T14:00:00.000Z'
...
    Project:
      required:
        - id
        - title
        - useCases
        - companyId
        - createdBy
        - archived
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        title:
          $ref: '#/components/schemas/ProjectTitle'
        useCases:
          $ref: '#/components/schemas/ProjectUseCases'
        companyId:
          $ref: '#/components/schemas/ProjectCompanyId'
        createdBy:
          $ref: '#/components/schemas/CreatedBy'
        fullCreatedBy:
          $ref: "#/components/schemas/User"
          description: Contains the user that created the project
        company:
          $ref: "#/components/schemas/Company"
          description: Contains the parent company
        archived:
          $ref: '#/components/schemas/ArchivedFields'
        associatedUsers:
          $ref: '#/components/schemas/AssociatedUserFields'
        createdAt:
          $ref: '#/components/schemas/CreatedAtFields'
        updatedAt:
          $ref: '#/components/schemas/UpdatedAtFields'
    PostProject:
      required:
        - title
        - companyId
      properties:
        title:
          $ref: '#/components/schemas/ProjectTitle'
        useCases:
          $ref: '#/components/schemas/ProjectUseCases'
        companyId:
          $ref: '#/components/schemas/ProjectCompanyId'
        associatedUsers:
          $ref: '#/components/schemas/AssociatedUserFields'
    PutProject:
      properties:
        title:
          $ref: '#/components/schemas/ProjectTitle'
        useCases:
          $ref: '#/components/schemas/ProjectUseCases'
        companyId:
          $ref: '#/components/schemas/ProjectCompanyId'
        associatedUsers:
          $ref: '#/components/schemas/AssociatedUserFields'
    SuccessProject:
      required:
        - project
      properties:
        project:
          type: object
          description: Contains the project that was successfully created/updated
          required:
            - id
            - title
            - useCases
            - companyId
            - createdBy
            - archived
          properties:
            id:
              $ref: '#/components/schemas/IdFields'
            title:
              $ref: '#/components/schemas/ProjectTitle'
            useCases:
              $ref: '#/components/schemas/ProjectUseCases'
            companyId:
              $ref: '#/components/schemas/ProjectCompanyId'
            createdBy:
              $ref: '#/components/schemas/CreatedBy'
            archived:
              $ref: '#/components/schemas/ArchivedFields'
            associatedUsers:
              $ref: '#/components/schemas/AssociatedUserFields'
            createdAt:
              $ref: '#/components/schemas/CreatedAtFields'
            updatedAt:
              $ref: '#/components/schemas/UpdatedAtFields'
...
    NotFound:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Found"
    NotAuthorized:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Authorized"